major update
ensure that dataframe export is working, shifted all of the IDs by 3
so that the first scientist born has unique_id of 1

printing out tuples instead of numpy arrays in dataframe is the magic
trick!

fixed logical errors in step, such as moving datacollector() after the
step() rather than before it.